Change the implementation of the type Latex.t using lists with fast concatenations.
Instead of using a constructor Concat of t list with arbitrary nesting, we use a single toplevel Clist.t. The direct result is that recursive functions over the type Latex.t have a much lower stack-depth (they are typically not tail-recursive, though, because the Clist.t can actually be nested through other constructions like "Command" or "Mode").
To lower the stack depth even more, we should try to make more recursive call tail.

My early tests seem to show a consistent but mostly insignificant improvement in execution speed of melt-generated executable. It is still to be verified that this new version can bring real benefits.

On a side note, it is worth noticing that this new implementation gives raise to almost identical code in latex.ml. Hence the cost of adopting it would be fairly low.