Removed support for tex_dependencies in mlpost_on.ml
It never worked, and it actually caused an interesting bug. Indeed, as the filename gensym relies on a reference, it would cause the fixpoint computation of latex to diverge when an Mlpost picture depended on a variable (in the sense of the Latex module) because tex_dependencies would be different at each iteration.
A further fix will be to change the gensym to use an actual Latex-module variable.