Removed as many side effects as possible wrt mlpost.
Mlpost integration used to emit files and call a reference during while building a latex document. This sort of things interact very badly with the fixed point computation. Instead we now use a variable as a gensym, and we delay calls to emit until Melt.emit is called.

In addition, Mlpost.Defaults.set_prelude is called during emit.